initial commit
You are given an integer array prices where prices[i] is the price of a given stock on the ith day.

On each day, you may decide to buy and/or sell the stock. You can only hold at most one share of the stock at any time. However, you can buy it then immediately sell it on the same day.

Find and return the maximum profit you can achieve.

 

Example 1:

Input: prices = [7,1,5,3,6,4]
Output: 7
Explanation: Buy on day 2 (price = 1) and sell on day 3 (price = 5), profit = 5-1 = 4.
Then buy on day 4 (price = 3) and sell on day 5 (price = 6), profit = 6-3 = 3.
Total profit is 4 + 3 = 7.

Example 2:

Input: prices = [1,2,3,4,5]
Output: 4
Explanation: Buy on day 1 (price = 1) and sell on day 5 (price = 5), profit = 5-1 = 4.
Total profit is 4.

Example 3:

Input: prices = [7,6,4,3,1]
Output: 0
Explanation: There is no way to make a positive profit, so we never buy the stock to achieve the maximum profit of 0.

 

Constraints:

    1 <= prices.length <= 3 * 104
    0 <= prices[i] <= 104

This solution just uses simple approach of trading - buy low, sell high. Think about how you could maximize your profits if you knew prices of stocks on each day with only 1 stock allowed at hand? You would of-course buy on the day where price was lowest in the local range and sell when it was highest in local range and perform as many trades as you can since you are allowed only 1 stock.

It's always better to consider every low & highs and trade between them and not skip any low/highs or not just consider the global minimum and maximum because it is always equal or more profitable to buy and sell at every local valleys (lowest) and peaks (highest) and trade multiple times in between if there are multiple valleys and peaks. This is best illustrated by LeetCode's official solution diagram (allow me to borrow it for educational purpose 😅) -

	We can see that we have choices to either perform two trades -

   1. Buy at valley2, sell at peak3 and again buy at valley4, sell at peak5.
    Profit = A+B
   2. Buy at valley2, sell at peak5. Profit = C

In any case, we have A + B > C because if A+B == C, then valley4 can't exist.
	We can see that we will buy at valley3 and sell at peak6.

As you will see later on in solution, this is simulated as buy3, sell 4+buy4, sell 5+buy5, sell 6 which is nothing but equivalent to buy3 and sell6

From above, we can see that we need to perform multiple trades buying at valley, selling at peak every time possible. Whenever we find that today's price is higher than yesterday, we know we should have bought yesterday and sold today. It can be implemented in a naive way as below -