initial commit
Given an array of integers temperatures represents the daily temperatures, return an array answer such that answer[i] is the number of days you have to wait after the ith day to get a warmer temperature. If there is no future day for which this is possible, keep answer[i] == 0 instead.

 

Example 1:

Input: temperatures = [73,74,75,71,69,72,76,73]
Output: [1,1,4,2,1,1,0,0]

Example 2:

Input: temperatures = [30,40,50,60]
Output: [1,1,1,0]

Example 3:

Input: temperatures = [30,60,90]
Output: [1,1,0]

 

Constraints:

    1 <= temperatures.length <= 105
    30 <= temperatures[i] <= 100


In this solution, we populate the result array in the reverse order.
For every index i:

    We have already populated the result array for all temperatures on the right side of i ([i+1, ..., len-1]).
    So we start another pointer, k from i+1, and start searching for a higher temperature. To perform this in an optimized manner (instead of checking all indexes), we follow the below approach:
        If temperatures[i] >= temperatures[k], then we have found a lower temperature. We can move our pointer k by the value stored at result[k]. We can safely do this because All temperatures between k andk + result[k] are smaller than temperatures[k]. By doing this, we skip checking all input temperatures and jump to the next possible higher temperature.
            Note: To avoid going into infinite loop, we also need to check if result[k] != 0. This can cause TLE, as it will never be able to come out of the while loop. This also helps us exit from this while loop if no higher temperature is found, as the rightmost value of the result will always be 0.
        If temperatures[i] < temperatures[k], then we have found the next higher temperature. We can set the result[i] to k - i. The difference in the pointers will give us the number of days to reach the higher temperature.

Time Complexity: O(N) --> Please refer to below for explanation.
Space Complexity: O(1) --> This solution uses constant space.
Where, N = Length of the input list.

Explanation of Time Complexity using an Exmaple
Lets say the worst case input is temperatures = [ 20, 15, 1, 2, 3, 21].
We can construct result upto [?, ?, 1, 1, 1, 0] in O(1) time for each index.
For 15 we will have to go through each temperature from 1 to 21. Thus it will visit each. result[1] = 4 ==> [?, 4, 1, 1, 1, 0]
For 20, we just have to look at 15 as it will directly jump the pointer k to k + result[k] = 1 + 4 = 5. This will cause the while loop to exit as result[5] is 0 since it's the last temperature in the input array. Thus, we get the result[0] in just two steps, without visiting all temperatures on the right.