initial commit
You are given two lists of closed intervals, firstList and secondList, where firstList[i] = [starti, endi] and secondList[j] = [startj, endj]. Each list of intervals is pairwise disjoint and in sorted order.

Return the intersection of these two interval lists.

A closed interval [a, b] (with a <= b) denotes the set of real numbers x with a <= x <= b.

The intersection of two closed intervals is a set of real numbers that are either empty or represented as a closed interval. For example, the intersection of [1, 3] and [2, 4] is [2, 3].

 

Example 1:
https://assets.leetcode.com/uploads/2019/01/30/interval1.png
Input: firstList = [[0,2],[5,10],[13,23],[24,25]], secondList = [[1,5],[8,12],[15,24],[25,26]]
Output: [[1,2],[5,5],[8,10],[15,23],[24,24],[25,25]]

Example 2:

Input: firstList = [[1,3],[5,9]], secondList = []
Output: []

Example 3:

Input: firstList = [], secondList = [[4,8],[10,12]]
Output: []

Example 4:

Input: firstList = [[1,7]], secondList = [[3,10]]
Output: [[3,7]]

 

Constraints:

    0 <= firstList.length, secondList.length <= 1000
    firstList.length + secondList.length >= 1
    0 <= starti < endi <= 109
    endi < starti+1
    0 <= startj < endj <= 109
    endj < startj+1


We just need to recognize the different possible intersections and then merge them accordingly. Let's see the possible types of intersections and how they can be generalized into single expression -

Intervals

	

Description

	

Generalized Expression

	

    Aend > Bstart and Astart > Bend
    No Intersection

	In every intersection, we have something like End1 >= Start2 and Start1 <= End2. We generalize it in terms of A and B as -


    Aend >= Bstart and Astart <= Bend

    Intersection = [Max(Astart, Bstart), Min(Aend, Bend)]

	

    Aend > Bstart and Astart < Bend
    Intersection = [Bstart, Aend]

	

    Bend > Astart and Bstart < Aend
    Intersection = [Astart, Bend]

	

    Aend > Bstart and Astart < Bend
    Intersection = [Astart, Bend]

	

    Aend = Bstart and Astart < Bend
    Intersection = [Bstart, Aend]

	

    Bend = Astart and Bstart < Aend
    Intersection = [Astart, Bend]

Now, we know the condition when an intersection will occur. We just need to figure out how we are going to traverse through the given list.

We begin from the starting intervals of both the list. Once we compare the intervals, either an intersection will occur or not occur. Then for moving to the next comparison, we need to observe that lists are sorted and the interval whose end came first will never be a potential candidate for intersection with other intervals (because if didnt intersect with current interval of other group, it wont intersect with any further interval as well, and if it did intersect with current interval of other group, it intersected all the way till its endpoint).

So for traversal, we just maintain two-pointers and after each comparison increment the pointer of the list whose interval had endpoint ending first.