initial commit
The complement of an integer is the integer you get when you flip all the 0's to 1's and all the 1's to 0's in its binary representation.

    For example, The integer 5 is "101" in binary and its complement is "010" which is the integer 2.

Given an integer num, return its complement.

 

Example 1:

Input: num = 5
Output: 2
Explanation: The binary representation of 5 is 101 (no leading zero bits), and its complement is 010. So you need to output 2.

Example 2:

Input: num = 1
Output: 0
Explanation: The binary representation of 1 is 1 (no leading zero bits), and its complement is 0. So you need to output 0.

 

Constraints:

    1 <= num < 231


Idea:

    Flip all bits using negation operator (~) [read more here]
    Bitwise AND (&) with a bit mask of size n whose all bits are set, where n = number of bits in num
    Example:

num = 5
  num binary = 0...0101
 ~num binary = 1...1010

* Now we've flipped all the bits but it also flipped previously insignificant bits (0s), 
so to revert them back to 0s, we need to only keep the significant bits and turn off the insignificant bits
* This can be done by using mask of 1s having size equal to number of bits in num. This mask is (1 << nBits) - 1
* This is a standard mask in bit manipulation problems

~num binary & mask = (1...1010) & (0...0111) = 0...0010 [Ans]