initial commit
We define the usage of capitals in a word to be right when one of the following cases holds:

    All letters in this word are capitals, like "USA".
    All letters in this word are not capitals, like "leetcode".
    Only the first letter in this word is capital, like "Google".

Given a string word, return true if the usage of capitals in it is right.

 

Example 1:

Input: word = "USA"
Output: true

Example 2:

Input: word = "FlaG"
Output: false

 

Constraints:

    1 <= word.length <= 100
    word consists of lowercase and uppercase English letters.

    Let's look at the first 2 character, if we have first 2 character there are 4 ways of writing it :- UU, UL, LU, LL

    So, if we have UU like :- AB____ then remaing one has to be uppercase
    If we have UL like :- Ab_____ then remaining one has to be lowercase
    If we have LU like :- aB then straight away we know that it's invalid
    If we have LL like :- ab_____ then remaining one has to be lowercase

In last 3 cases Ab & aB & ab we can check from 2nd letter onwards everything should be lower. So, aB is invalid because the 2nd letter is capital
But if 1st and 2nd both are capital we will check from 3rd letter everyting should be upper in AB. So this is valid

So, in our solution we create 2 cases in 1st case everything has to be UpperCase from 3rd letter and in 2nd case everything has to be LowerCase from 2nd letter. If any of these condition invalidate we return false otherwise we return true

I hope you get the idea, Now let's code it up :

code each line explained : Similar for C++, Java, Python {Only Syntax Difference} approach same

{
        if(word.length() < 2) return true;// either if word is empty or word is just 1 character long, then it's always valid
        // isupper OR islower is built in function in C++, but all the language have but available in different syntax format
        // but if you don't want to use you can write word[0] >= 'A' ASCII value && word[1] <= 'Z'
        if(isupper(word[0]) && isupper(word[1])){ // creating our 1st condition, from 3rd letter onward everything should be capital
            for(int i = 2; i < word.length(); i++){
                if(islower(word[i])) return false;
            }
        }
        else{ // creating our 2nd condition, from 2nd letter onward everything should be small
            for(int i = 1; i < word.length(); i++){
                if(isupper(word[i])) return false;
            }
        }
        return true;