initial commit
Given an array of integers nums and an integer k, return the total number of continuous subarrays whose sum equals to k.

 

Example 1:

Input: nums = [1,1,1], k = 2
Output: 2
Example 2:

Input: nums = [1,2,3], k = 3
Output: 2
 

Constraints:

1 <= nums.length <= 2 * 104
-1000 <= nums[i] <= 1000
-107 <= k <= 107


Explaination :

First we will create HashMap where first Integer is for storing cumulative sum of subarrays and second Integer is for frequencies of that sum.

After that let 's put 0 into Hashmap as we seen it for 1 time at the beginning.

Afer that we will iterate through loop and we will add each element into sum.
So same sum can come again and again let's take an example.
arr = [1,2,3,-2,2, 3] ==> so here sum == 6 apperas at 2 times (1. For first 3 elements of sub_array ->[1,2,3] and 2. For first 5 elements of sub_array -> [1,2,3,-2,2] ) so we need to increment that count in our Hashmap.

Now we will check whether our HashMap contains key (sum-k) or not.( using arr_sums.containsKey(sum - k)) ==> And if it contains key which we required as (sum-k) then we will count it into our answer as a correct subarray which has sum equals to k.

After that, we have 2 cases when putting sum into HashMap which we will be using for next remaining iterations.

 		arr_sums.put(sum, arr_sums.getOrDefault(sum, 0) + 1);
a) If our sum is already exist in our HashMap, then we will get the frequencies of that sum and add 1 to it.
b) else our sum is not exist, then frequency will be 0 and we will add 1 into it.